backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
passlib[bcrypt] # For password hashing
alembic        # For database migrations
python-dotenv   # For environment variables (.env)
requests      # If you need to make HTTP requests
